Extract logging to new logging.py
